requirements.txt
tensorflow==2.13.0
streamlit==1.24.0
numpy==1.23.5
Pillow==9.5.0
pandas==2.0.3
